cycle-homophones and split-at−apostrophe
Hello,
I'm making my first contribution to the registry with these two plugins
that are meant to be used with my modded-grandjean system. I'll be making
another PR in the upcoming days to also add my grandjean system to the
registry.

Please let me know if there's anything I've missed in the process of
updating the registry.

The two current plugins can be found at:
https://pypi.org/project/plover-cycle-homophones/0.0.1/
https://pypi.org/project/plover-split-at-apostrophe/0.0.5/

Best regards,

Antoine